+++ clients/java/src/main/java/io/lakefs/clients/api/ServerConfiguration.java
@@ -39,10 +39,10 @@ public class ServerConfiguration {
             if (variables != null && variables.containsKey(name)) {
                 value = variables.get(name);
                 if (serverVariable.enumValues.size() > 0 && !serverVariable.enumValues.contains(value)) {
-                    throw new RuntimeException("The variable " + name + " in the server URL has invalid value " + value + ".");
+                    throw new IllegalArgumentException("The variable " + name + " in the server URL has invalid value " + value + ".");
                 }
             }
-            url = url.replaceAll("\\{" + name + "\\}", value);
+            url = url.replace("{" + name + "}", value);
         }
         return url;
     }