+++ clients/java/src/main/java/io/lakefs/clients/api/model/HookRunList.java
@@ -22,11 +22,30 @@ import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
 import io.lakefs.clients.api.model.HookRun;
 import io.lakefs.clients.api.model.Pagination;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonDeserializationContext;
+import com.google.gson.JsonDeserializer;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParseException;
+import com.google.gson.TypeAdapterFactory;
+import com.google.gson.reflect.TypeToken;
+
+import java.lang.reflect.Type;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.Set;
+
+import io.lakefs.clients.api.JSON;
 
 /**
  * HookRunList
@@ -39,8 +58,10 @@ public class HookRunList {
 
   public static final String SERIALIZED_NAME_RESULTS = "results";
   @SerializedName(SERIALIZED_NAME_RESULTS)
-  private List<HookRun> results = new ArrayList<HookRun>();
+  private List<HookRun> results = new ArrayList<>();
 
+  public HookRunList() {
+  }
 
   public HookRunList pagination(Pagination pagination) {
     
@@ -53,7 +74,6 @@ public class HookRunList {
    * @return pagination
   **/
   @javax.annotation.Nonnull
-  @ApiModelProperty(required = true, value = "")
 
   public Pagination getPagination() {
     return pagination;
@@ -81,7 +101,6 @@ public class HookRunList {
    * @return results
   **/
   @javax.annotation.Nonnull
-  @ApiModelProperty(required = true, value = "")
 
   public List<HookRun> getResults() {
     return results;
@@ -91,6 +110,7 @@ public class HookRunList {
   public void setResults(List<HookRun> results) {
     this.results = results;
   }
+
 
 
   @Override
@@ -132,5 +152,110 @@ public class HookRunList {
     return o.toString().replace("\n", "\n    ");
   }
 
+
+  public static HashSet<String> openapiFields;
+  public static HashSet<String> openapiRequiredFields;
+
+  static {
+    // a set of all properties/fields (JSON key names)
+    openapiFields = new HashSet<String>();
+    openapiFields.add("pagination");
+    openapiFields.add("results");
+
+    // a set of required properties/fields (JSON key names)
+    openapiRequiredFields = new HashSet<String>();
+    openapiRequiredFields.add("pagination");
+    openapiRequiredFields.add("results");
+  }
+
+ /**
+  * Validates the JSON Object and throws an exception if issues found
+  *
+  * @param jsonObj JSON Object
+  * @throws IOException if the JSON Object is invalid with respect to HookRunList
+  */
+  public static void validateJsonObject(JsonObject jsonObj) throws IOException {
+      if (jsonObj == null) {
+        if (!HookRunList.openapiRequiredFields.isEmpty()) { // has required fields but JSON object is null
+          throw new IllegalArgumentException(String.format("The required field(s) %s in HookRunList is not found in the empty JSON string", HookRunList.openapiRequiredFields.toString()));
+        }
+      }
+
+      Set<Entry<String, JsonElement>> entries = jsonObj.entrySet();
+      // check to see if the JSON string contains additional fields
+      for (Entry<String, JsonElement> entry : entries) {
+        if (!HookRunList.openapiFields.contains(entry.getKey())) {
+          throw new IllegalArgumentException(String.format("The field `%s` in the JSON string is not defined in the `HookRunList` properties. JSON: %s", entry.getKey(), jsonObj.toString()));
+        }
+      }
+
+      // check to make sure all required properties/fields are present in the JSON string
+      for (String requiredField : HookRunList.openapiRequiredFields) {
+        if (jsonObj.get(requiredField) == null) {
+          throw new IllegalArgumentException(String.format("The required field `%s` is not found in the JSON string: %s", requiredField, jsonObj.toString()));
+        }
+      }
+      // validate the required field `pagination`
+      Pagination.validateJsonObject(jsonObj.getAsJsonObject("pagination"));
+      // ensure the json data is an array
+      if (!jsonObj.get("results").isJsonArray()) {
+        throw new IllegalArgumentException(String.format("Expected the field `results` to be an array in the JSON string but got `%s`", jsonObj.get("results").toString()));
+      }
+
+      JsonArray jsonArrayresults = jsonObj.getAsJsonArray("results");
+      // validate the required field `results` (array)
+      for (int i = 0; i < jsonArrayresults.size(); i++) {
+        HookRun.validateJsonObject(jsonArrayresults.get(i).getAsJsonObject());
+      };
+  }
+
+  public static class CustomTypeAdapterFactory implements TypeAdapterFactory {
+    @SuppressWarnings("unchecked")
+    @Override
+    public <T> TypeAdapter<T> create(Gson gson, TypeToken<T> type) {
+       if (!HookRunList.class.isAssignableFrom(type.getRawType())) {
+         return null; // this class only serializes 'HookRunList' and its subtypes
+       }
+       final TypeAdapter<JsonElement> elementAdapter = gson.getAdapter(JsonElement.class);
+       final TypeAdapter<HookRunList> thisAdapter
+                        = gson.getDelegateAdapter(this, TypeToken.get(HookRunList.class));
+
+       return (TypeAdapter<T>) new TypeAdapter<HookRunList>() {
+           @Override
+           public void write(JsonWriter out, HookRunList value) throws IOException {
+             JsonObject obj = thisAdapter.toJsonTree(value).getAsJsonObject();
+             elementAdapter.write(out, obj);
+           }
+
+           @Override
+           public HookRunList read(JsonReader in) throws IOException {
+             JsonObject jsonObj = elementAdapter.read(in).getAsJsonObject();
+             validateJsonObject(jsonObj);
+             return thisAdapter.fromJsonTree(jsonObj);
+           }
+
+       }.nullSafe();
+    }
+  }
+
+ /**
+  * Create an instance of HookRunList given an JSON string
+  *
+  * @param jsonString JSON string
+  * @return An instance of HookRunList
+  * @throws IOException if the JSON string is invalid with respect to HookRunList
+  */
+  public static HookRunList fromJson(String jsonString) throws IOException {
+    return JSON.getGson().fromJson(jsonString, HookRunList.class);
+  }
+
+ /**
+  * Convert an instance of HookRunList to an JSON string
+  *
+  * @return JSON string
+  */
+  public String toJson() {
+    return JSON.getGson().toJson(this);
+  }
 }