+++ clients/java/src/main/java/io/lakefs/clients/api/model/RepositoryCreation.java
@@ -20,9 +20,28 @@ import com.google.gson.annotations.JsonAdapter;
 import com.google.gson.annotations.SerializedName;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
 import java.io.IOException;
+
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonDeserializationContext;
+import com.google.gson.JsonDeserializer;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParseException;
+import com.google.gson.TypeAdapterFactory;
+import com.google.gson.reflect.TypeToken;
+
+import java.lang.reflect.Type;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.Set;
+
+import io.lakefs.clients.api.JSON;
 
 /**
  * RepositoryCreation
@@ -41,6 +60,8 @@ public class RepositoryCreation {
   @SerializedName(SERIALIZED_NAME_DEFAULT_BRANCH)
   private String defaultBranch;
 
+  public RepositoryCreation() {
+  }
 
   public RepositoryCreation name(String name) {
     
@@ -53,7 +74,6 @@ public class RepositoryCreation {
    * @return name
   **/
   @javax.annotation.Nonnull
-  @ApiModelProperty(required = true, value = "")
 
   public String getName() {
     return name;
@@ -76,7 +96,6 @@ public class RepositoryCreation {
    * @return storageNamespace
   **/
   @javax.annotation.Nonnull
-  @ApiModelProperty(example = "s3://example-bucket/", required = true, value = "Filesystem URI to store the underlying data in (e.g. \"s3://my-bucket/some/path/\")")
 
   public String getStorageNamespace() {
     return storageNamespace;
@@ -99,7 +118,6 @@ public class RepositoryCreation {
    * @return defaultBranch
   **/
   @javax.annotation.Nullable
-  @ApiModelProperty(example = "main", value = "")
 
   public String getDefaultBranch() {
     return defaultBranch;
@@ -109,6 +127,7 @@ public class RepositoryCreation {
   public void setDefaultBranch(String defaultBranch) {
     this.defaultBranch = defaultBranch;
   }
+
 
 
   @Override
@@ -152,5 +171,108 @@ public class RepositoryCreation {
     return o.toString().replace("\n", "\n    ");
   }
 
+
+  public static HashSet<String> openapiFields;
+  public static HashSet<String> openapiRequiredFields;
+
+  static {
+    // a set of all properties/fields (JSON key names)
+    openapiFields = new HashSet<String>();
+    openapiFields.add("name");
+    openapiFields.add("storage_namespace");
+    openapiFields.add("default_branch");
+
+    // a set of required properties/fields (JSON key names)
+    openapiRequiredFields = new HashSet<String>();
+    openapiRequiredFields.add("name");
+    openapiRequiredFields.add("storage_namespace");
+  }
+
+ /**
+  * Validates the JSON Object and throws an exception if issues found
+  *
+  * @param jsonObj JSON Object
+  * @throws IOException if the JSON Object is invalid with respect to RepositoryCreation
+  */
+  public static void validateJsonObject(JsonObject jsonObj) throws IOException {
+      if (jsonObj == null) {
+        if (!RepositoryCreation.openapiRequiredFields.isEmpty()) { // has required fields but JSON object is null
+          throw new IllegalArgumentException(String.format("The required field(s) %s in RepositoryCreation is not found in the empty JSON string", RepositoryCreation.openapiRequiredFields.toString()));
+        }
+      }
+
+      Set<Entry<String, JsonElement>> entries = jsonObj.entrySet();
+      // check to see if the JSON string contains additional fields
+      for (Entry<String, JsonElement> entry : entries) {
+        if (!RepositoryCreation.openapiFields.contains(entry.getKey())) {
+          throw new IllegalArgumentException(String.format("The field `%s` in the JSON string is not defined in the `RepositoryCreation` properties. JSON: %s", entry.getKey(), jsonObj.toString()));
+        }
+      }
+
+      // check to make sure all required properties/fields are present in the JSON string
+      for (String requiredField : RepositoryCreation.openapiRequiredFields) {
+        if (jsonObj.get(requiredField) == null) {
+          throw new IllegalArgumentException(String.format("The required field `%s` is not found in the JSON string: %s", requiredField, jsonObj.toString()));
+        }
+      }
+      if (!jsonObj.get("name").isJsonPrimitive()) {
+        throw new IllegalArgumentException(String.format("Expected the field `name` to be a primitive type in the JSON string but got `%s`", jsonObj.get("name").toString()));
+      }
+      if (!jsonObj.get("storage_namespace").isJsonPrimitive()) {
+        throw new IllegalArgumentException(String.format("Expected the field `storage_namespace` to be a primitive type in the JSON string but got `%s`", jsonObj.get("storage_namespace").toString()));
+      }
+      if ((jsonObj.get("default_branch") != null && !jsonObj.get("default_branch").isJsonNull()) && !jsonObj.get("default_branch").isJsonPrimitive()) {
+        throw new IllegalArgumentException(String.format("Expected the field `default_branch` to be a primitive type in the JSON string but got `%s`", jsonObj.get("default_branch").toString()));
+      }
+  }
+
+  public static class CustomTypeAdapterFactory implements TypeAdapterFactory {
+    @SuppressWarnings("unchecked")
+    @Override
+    public <T> TypeAdapter<T> create(Gson gson, TypeToken<T> type) {
+       if (!RepositoryCreation.class.isAssignableFrom(type.getRawType())) {
+         return null; // this class only serializes 'RepositoryCreation' and its subtypes
+       }
+       final TypeAdapter<JsonElement> elementAdapter = gson.getAdapter(JsonElement.class);
+       final TypeAdapter<RepositoryCreation> thisAdapter
+                        = gson.getDelegateAdapter(this, TypeToken.get(RepositoryCreation.class));
+
+       return (TypeAdapter<T>) new TypeAdapter<RepositoryCreation>() {
+           @Override
+           public void write(JsonWriter out, RepositoryCreation value) throws IOException {
+             JsonObject obj = thisAdapter.toJsonTree(value).getAsJsonObject();
+             elementAdapter.write(out, obj);
+           }
+
+           @Override
+           public RepositoryCreation read(JsonReader in) throws IOException {
+             JsonObject jsonObj = elementAdapter.read(in).getAsJsonObject();
+             validateJsonObject(jsonObj);
+             return thisAdapter.fromJsonTree(jsonObj);
+           }
+
+       }.nullSafe();
+    }
+  }
+
+ /**
+  * Create an instance of RepositoryCreation given an JSON string
+  *
+  * @param jsonString JSON string
+  * @return An instance of RepositoryCreation
+  * @throws IOException if the JSON string is invalid with respect to RepositoryCreation
+  */
+  public static RepositoryCreation fromJson(String jsonString) throws IOException {
+    return JSON.getGson().fromJson(jsonString, RepositoryCreation.class);
+  }
+
+ /**
+  * Convert an instance of RepositoryCreation to an JSON string
+  *
+  * @return JSON string
+  */
+  public String toJson() {
+    return JSON.getGson().toJson(this);
+  }
 }