+++ clients/java/src/main/java/io/lakefs/clients/api/model/VersionConfig.java
@@ -20,9 +20,28 @@ import com.google.gson.annotations.JsonAdapter;
 import com.google.gson.annotations.SerializedName;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
 import java.io.IOException;
+
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonDeserializationContext;
+import com.google.gson.JsonDeserializer;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParseException;
+import com.google.gson.TypeAdapterFactory;
+import com.google.gson.reflect.TypeToken;
+
+import java.lang.reflect.Type;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.Set;
+
+import io.lakefs.clients.api.JSON;
 
 /**
  * VersionConfig
@@ -41,6 +60,8 @@ public class VersionConfig {
   @SerializedName(SERIALIZED_NAME_UPGRADE_URL)
   private String upgradeUrl;
 
+  public VersionConfig() {
+  }
 
   public VersionConfig version(String version) {
     
@@ -53,7 +74,6 @@ public class VersionConfig {
    * @return version
   **/
   @javax.annotation.Nullable
-  @ApiModelProperty(value = "")
 
   public String getVersion() {
     return version;
@@ -76,7 +96,6 @@ public class VersionConfig {
    * @return upgradeRecommended
   **/
   @javax.annotation.Nullable
-  @ApiModelProperty(value = "")
 
   public Boolean getUpgradeRecommended() {
     return upgradeRecommended;
@@ -99,7 +118,6 @@ public class VersionConfig {
    * @return upgradeUrl
   **/
   @javax.annotation.Nullable
-  @ApiModelProperty(value = "")
 
   public String getUpgradeUrl() {
     return upgradeUrl;
@@ -109,6 +127,7 @@ public class VersionConfig {
   public void setUpgradeUrl(String upgradeUrl) {
     this.upgradeUrl = upgradeUrl;
   }
+
 
 
   @Override
@@ -152,5 +171,96 @@ public class VersionConfig {
     return o.toString().replace("\n", "\n    ");
   }
 
+
+  public static HashSet<String> openapiFields;
+  public static HashSet<String> openapiRequiredFields;
+
+  static {
+    // a set of all properties/fields (JSON key names)
+    openapiFields = new HashSet<String>();
+    openapiFields.add("version");
+    openapiFields.add("upgrade_recommended");
+    openapiFields.add("upgrade_url");
+
+    // a set of required properties/fields (JSON key names)
+    openapiRequiredFields = new HashSet<String>();
+  }
+
+ /**
+  * Validates the JSON Object and throws an exception if issues found
+  *
+  * @param jsonObj JSON Object
+  * @throws IOException if the JSON Object is invalid with respect to VersionConfig
+  */
+  public static void validateJsonObject(JsonObject jsonObj) throws IOException {
+      if (jsonObj == null) {
+        if (!VersionConfig.openapiRequiredFields.isEmpty()) { // has required fields but JSON object is null
+          throw new IllegalArgumentException(String.format("The required field(s) %s in VersionConfig is not found in the empty JSON string", VersionConfig.openapiRequiredFields.toString()));
+        }
+      }
+
+      Set<Entry<String, JsonElement>> entries = jsonObj.entrySet();
+      // check to see if the JSON string contains additional fields
+      for (Entry<String, JsonElement> entry : entries) {
+        if (!VersionConfig.openapiFields.contains(entry.getKey())) {
+          throw new IllegalArgumentException(String.format("The field `%s` in the JSON string is not defined in the `VersionConfig` properties. JSON: %s", entry.getKey(), jsonObj.toString()));
+        }
+      }
+      if ((jsonObj.get("version") != null && !jsonObj.get("version").isJsonNull()) && !jsonObj.get("version").isJsonPrimitive()) {
+        throw new IllegalArgumentException(String.format("Expected the field `version` to be a primitive type in the JSON string but got `%s`", jsonObj.get("version").toString()));
+      }
+      if ((jsonObj.get("upgrade_url") != null && !jsonObj.get("upgrade_url").isJsonNull()) && !jsonObj.get("upgrade_url").isJsonPrimitive()) {
+        throw new IllegalArgumentException(String.format("Expected the field `upgrade_url` to be a primitive type in the JSON string but got `%s`", jsonObj.get("upgrade_url").toString()));
+      }
+  }
+
+  public static class CustomTypeAdapterFactory implements TypeAdapterFactory {
+    @SuppressWarnings("unchecked")
+    @Override
+    public <T> TypeAdapter<T> create(Gson gson, TypeToken<T> type) {
+       if (!VersionConfig.class.isAssignableFrom(type.getRawType())) {
+         return null; // this class only serializes 'VersionConfig' and its subtypes
+       }
+       final TypeAdapter<JsonElement> elementAdapter = gson.getAdapter(JsonElement.class);
+       final TypeAdapter<VersionConfig> thisAdapter
+                        = gson.getDelegateAdapter(this, TypeToken.get(VersionConfig.class));
+
+       return (TypeAdapter<T>) new TypeAdapter<VersionConfig>() {
+           @Override
+           public void write(JsonWriter out, VersionConfig value) throws IOException {
+             JsonObject obj = thisAdapter.toJsonTree(value).getAsJsonObject();
+             elementAdapter.write(out, obj);
+           }
+
+           @Override
+           public VersionConfig read(JsonReader in) throws IOException {
+             JsonObject jsonObj = elementAdapter.read(in).getAsJsonObject();
+             validateJsonObject(jsonObj);
+             return thisAdapter.fromJsonTree(jsonObj);
+           }
+
+       }.nullSafe();
+    }
+  }
+
+ /**
+  * Create an instance of VersionConfig given an JSON string
+  *
+  * @param jsonString JSON string
+  * @return An instance of VersionConfig
+  * @throws IOException if the JSON string is invalid with respect to VersionConfig
+  */
+  public static VersionConfig fromJson(String jsonString) throws IOException {
+    return JSON.getGson().fromJson(jsonString, VersionConfig.class);
+  }
+
+ /**
+  * Convert an instance of VersionConfig to an JSON string
+  *
+  * @return JSON string
+  */
+  public String toJson() {
+    return JSON.getGson().toJson(this);
+  }
 }