+++ clients/java/src/test/java/io/lakefs/clients/api/OtfDiffApiTest.java
@@ -16,8 +16,8 @@ package io.lakefs.clients.api;
 import io.lakefs.clients.api.ApiException;
 import io.lakefs.clients.api.model.Error;
 import io.lakefs.clients.api.model.OtfDiffList;
-import org.junit.Test;
-import org.junit.Ignore;
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -27,19 +27,15 @@ import java.util.Map;
 /**
  * API tests for OtfDiffApi
  */
-@Ignore
+@Disabled
 public class OtfDiffApiTest {
 
     private final OtfDiffApi api = new OtfDiffApi();
 
-    
     /**
      * perform otf diff
      *
-     * 
-     *
-     * @throws ApiException
-     *          if the Api call fails
+     * @throws ApiException if the Api call fails
      */
     @Test
     public void otfDiffTest() throws ApiException {
@@ -48,8 +44,8 @@ public class OtfDiffApiTest {
         String rightRef = null;
         String tablePath = null;
         String type = null;
-                OtfDiffList response = api.otfDiff(repository, leftRef, rightRef, tablePath, type);
+        OtfDiffList response = api.otfDiff(repository, leftRef, rightRef, tablePath, type);
         // TODO: test validations
     }
-    
+
 }