+++ clients/python/lakefs_client/apis/paths/auth_groups.py
@@ -1,0 +1,9 @@
+from lakefs_client.paths.auth_groups.get import ApiForget
+from lakefs_client.paths.auth_groups.post import ApiForpost
+
+
+class AuthGroups(
+    ApiForget,
+    ApiForpost,
+):
+    pass