+++ clients/python/lakefs_client/apis/paths/config_garbage_collection.py
@@ -1,0 +1,7 @@
+from lakefs_client.paths.config_garbage_collection.get import ApiForget
+
+
+class ConfigGarbageCollection(
+    ApiForget,
+):
+    pass