+++ clients/python/lakefs_client/apis/paths/config_storage.py
@@ -1,0 +1,7 @@
+from lakefs_client.paths.config_storage.get import ApiForget
+
+
+class ConfigStorage(
+    ApiForget,
+):
+    pass