+++ clients/python/lakefs_client/apis/paths/repositories_repository_branches.py
@@ -1,0 +1,9 @@
+from lakefs_client.paths.repositories_repository_branches.get import ApiForget
+from lakefs_client.paths.repositories_repository_branches.post import ApiForpost
+
+
+class RepositoriesRepositoryBranches(
+    ApiForget,
+    ApiForpost,
+):
+    pass