+++ clients/python/lakefs_client/apis/paths/repositories_repository_branches_branch_staging_backing.py
@@ -1,0 +1,9 @@
+from lakefs_client.paths.repositories_repository_branches_branch_staging_backing.get import ApiForget
+from lakefs_client.paths.repositories_repository_branches_branch_staging_backing.put import ApiForput
+
+
+class RepositoriesRepositoryBranchesBranchStagingBacking(
+    ApiForget,
+    ApiForput,
+):
+    pass