+++ clients/python/lakefs_client/apis/paths/repositories_repository_gc_rules.py
@@ -1,0 +1,11 @@
+from lakefs_client.paths.repositories_repository_gc_rules.get import ApiForget
+from lakefs_client.paths.repositories_repository_gc_rules.post import ApiForpost
+from lakefs_client.paths.repositories_repository_gc_rules.delete import ApiFordelete
+
+
+class RepositoriesRepositoryGcRules(
+    ApiForget,
+    ApiForpost,
+    ApiFordelete,
+):
+    pass