+++ clients/python/lakefs_client/apis/paths/repositories_repository_refs_ref_commits.py
@@ -1,0 +1,7 @@
+from lakefs_client.paths.repositories_repository_refs_ref_commits.get import ApiForget
+
+
+class RepositoriesRepositoryRefsRefCommits(
+    ApiForget,
+):
+    pass