+++ clients/python/lakefs_client/apis/paths/repositories_repository_refs_ref_objects_underlying_properties.py
@@ -1,0 +1,7 @@
+from lakefs_client.paths.repositories_repository_refs_ref_objects_underlying_properties.get import ApiForget
+
+
+class RepositoriesRepositoryRefsRefObjectsUnderlyingProperties(
+    ApiForget,
+):
+    pass